browser_demo/pyproject.toml
[project]
name = "browser-demo"
version = "0.1.0"
description = "Playwright-based web clipper — fetch URLs, convert to markdown, store and search locally."
requires-python = ">=3.13"
license = { text = "MIT" }

dependencies = [
  "playwright>=1.40",
  "beautifulsoup4>=4.12",
  "markdownify>=0.13",
  "pydantic>=2.0",
  "typer>=0.12",
  "rich>=13",
  "pyyaml>=6",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.5.0",
]

[project.scripts]
remora-clip = "browser_demo.cli:app"

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/browser_demo"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "integration: requires Playwright browsers to be installed",
]

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008", "E501"]
